Surgen varios problemas a raiz de las validaciones. No deja editar productos y no valida el archivo img de los productos

+++ Proyecto/app/Http/Controllers/ProductController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use App\Http\Requests\ProductCreateRequest;
+use App\Http\Requests\ProductEditRequest;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -14,7 +17,8 @@ class ProductController extends Controller
      */
     public function index()
     {
-       $products = Product::paginate(5);
+        abort_if(Gate::denies('product_index'), 403);
+        $products = Product::paginate(5);
         return view('Productviews.index', compact('products'));
     }
 
@@ -25,6 +29,7 @@ class ProductController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('product_create'), 403);
         return view('Productviews.create');
     }
 
@@ -34,7 +39,7 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ProductCreateRequest $request)
     {
 
         //  $validator = validator::make($request->all(),[
@@ -79,6 +84,7 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        abort_if(Gate::denies('product_show'), 403);
       //  $product = Product::findOrfail($id);
 
          return view('Productviews.show', compact ('product'));
@@ -94,6 +100,7 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        abort_if(Gate::denies('product_edit'), 403);
         return view('Productviews.edit',compact('product'));
     }
 
@@ -104,7 +111,7 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductEditRequest $request, Product $product)
 
     {
         if($request['imagenes']){
@@ -130,8 +137,9 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-     $product->delete();
-     return back()->with('succes','usuario eliminado correctamente');
+        abort_if(Gate::denies('product_destroy'), 403);
+        $product->delete();
+        return back()->with('succes','usuario eliminado correctamente');
     }
